# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FlightMovement API Documentation",
        "description": "API documentation for FlightMovement",
        "contact": {
            "email": "dev@camranh.aero"
        },
        "version": "1.0.0"
    },
    "servers": [
        {

            "url": "http://localhost:8000/api",
            "description": "Rest API"
        }
    ],
    "paths": {
...
        "/airlines": {
            "get": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Get all active airlines",
                "description": "Retrieve all active airlines ordered by name",
                "operationId": "20835eb6981d4fa2332730fb19c119f6",
                "responses": {
                    "200": {
                        "description": "Airlines retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "AirlineId": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "Name": {
                                                        "type": "string",
                                                        "example": "Vietnam Airlines"
                                                    },
                                                    "IataCode": {
                                                        "type": "string",
                                                        "example": "VN"
                                                    },
                                                    "IcaoCode": {
                                                        "type": "string",
                                                        "example": "HVN"
                                                    },
                                                    "Handler": {
                                                        "type": "string",
                                                        "example": "SAGS"
                                                    },
                                                    "Status": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 15
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Create new airline",
                "description": "Create a new airline record",
                "operationId": "fddb1199b1630d20aeac780866549158",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "AirlineId",
                                    "Name"
                                ],
                                "properties": {
                                    "AirlineId": {
                                        "type": "integer",
                                        "example": 25
                                    },
                                    "Name": {
                                        "type": "string",
                                        "example": "New Airlines"
                                    },
                                    "IataCode": {
                                        "type": "string",
                                        "example": "NA"
                                    },
                                    "IcaoCode": {
                                        "type": "string",
                                        "example": "NEW"
                                    },
                                    "Handler": {
                                        "type": "string",
                                        "example": "SAGS"
                                    },
                                    "Status": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    },
                                    "IsKorea": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Airline created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Airline created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/airlines/iata/{iataCode}": {
            "get": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Get airline by IATA code",
                "description": "Retrieve airline information by IATA code",
                "operationId": "2cfb08ffef7f8d883e12f7fe85254f59",
                "parameters": [
                    {
                        "name": "iataCode",
                        "in": "path",
                        "description": "IATA code (e.g., VN, VJ)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "VN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Airline found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "AirlineId": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "Name": {
                                                    "type": "string",
                                                    "example": "Vietnam Airlines"
                                                },
                                                "IataCode": {
                                                    "type": "string",
                                                    "example": "VN"
                                                },
                                                "IcaoCode": {
                                                    "type": "string",
                                                    "example": "HVN"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Airline not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Airline not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/airlines/{airlineId}": {
            "put": {
                "tags": [
                    "Airlines"
                ],
                "summary": "Update airline",
                "description": "Update an existing airline record",
                "operationId": "4a4ef9c9eb0ffca18a046e2e0504dcea",
                "parameters": [
                    {
                        "name": "airlineId",
                        "in": "path",
                        "description": "Airline ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Name": {
                                        "type": "string",
                                        "example": "Updated Airline Name"
                                    },
                                    "IataCode": {
                                        "type": "string",
                                        "example": "UA"
                                    },
                                    "IcaoCode": {
                                        "type": "string",
                                        "example": "UPD"
                                    },
                                    "Status": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Airline updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Airline updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Airline not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Airline not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}